+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [x] Compile parent .less files when child file changes
 - [x] Context menu command to enable compiler for project
 - [x] Adornment to toggle compilation per solution
+- [x] Command to compile less files in all enabled projects in solution
 - [ ] Compile only parent files and not all files
 
 Features that have a checkmark are complete and available for